notebooks/azure_ml_experiment_tracking.py
"""
Azure Machine Learning — Experiment Tracking for PhiSphere AI

This script demonstrates how PhiSphere AI lab sessions can be tracked and
versioned using Azure Machine Learning. It registers datasets exported from
PhiSphere sessions, logs experiment metrics, and optionally trains a simple
model — all using the Azure ML SDK v2.

Use case for hackathon judges:
- Shows breadth of Azure services (Azure ML in addition to Cognitive Services)
- Demonstrates experiment reproducibility and MLOps readiness
- Connects PhiSphere's session export to Azure ML's dataset + experiment tracking

Prerequisites:
    pip install azure-ai-ml azure-identity pandas scikit-learn

Environment variables required:
    AZURE_SUBSCRIPTION_ID      — Your Azure subscription
    AZURE_RESOURCE_GROUP       — Resource group containing the Azure ML workspace
    AZURE_ML_WORKSPACE_NAME    — Azure ML workspace name

Usage:
    python notebooks/azure_ml_experiment_tracking.py --csv exported_session.csv
"""

import argparse
import os
import json
import sys
from pathlib import Path

def main():
    parser = argparse.ArgumentParser(description="PhiSphere AI — Azure ML Experiment Tracking")
    parser.add_argument("--csv", type=str, help="Path to exported CSV from a PhiSphere lab session")
    parser.add_argument("--dry-run", action="store_true", help="Run without Azure ML connection (local demo)")
    args = parser.parse_args()

    csv_path = args.csv
    dry_run = args.dry_run

    if not csv_path:
        print("No --csv provided. Using built-in sample data for demo.")
        csv_path = None

    import pandas as pd
    import numpy as np

    if csv_path and Path(csv_path).exists():
        df = pd.read_csv(csv_path)
        print(f"Loaded {len(df)} rows from {csv_path}")
    else:
        np.random.seed(42)
        n = 100
        df = pd.DataFrame({
            "time_hours": np.arange(0, n * 6, 6),
            "temperature_c": np.random.normal(24, 1.5, n).round(1),
            "humidity_pct": np.random.normal(68, 3, n).round(1),
            "co2_ppm": np.random.normal(415, 12, n).round(0),
            "growth_mm": np.cumsum(np.random.exponential(0.3, n)).round(1),
        })
        print(f"Generated synthetic plant growth dataset: {len(df)} rows")

    numeric_cols = df.select_dtypes(include=[np.number]).columns.tolist()
    stats = {}
    for col in numeric_cols:
        stats[col] = {
            "min": float(df[col].min()),
            "max": float(df[col].max()),
            "mean": float(df[col].mean()),
            "std": float(df[col].std()),
        }

    print("\n--- Dataset Statistics ---")
    for col, s in stats.items():
        print(f"  {col}: min={s['min']:.2f}, max={s['max']:.2f}, mean={s['mean']:.2f}, std={s['std']:.2f}")

    model_metrics = {}
    if len(numeric_cols) >= 2:
        from sklearn.linear_model import LinearRegression
        from sklearn.model_selection import train_test_split
        from sklearn.metrics import mean_squared_error, r2_score

        target_col = numeric_cols[-1]
        feature_cols = numeric_cols[:-1]

        X = df[feature_cols].fillna(0)
        y = df[target_col].fillna(0)

        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

        model = LinearRegression()
        model.fit(X_train, y_train)
        y_pred = model.predict(X_test)

        model_metrics = {
            "r2_score": round(float(r2_score(y_test, y_pred)), 4),
            "rmse": round(float(np.sqrt(mean_squared_error(y_test, y_pred))), 4),
            "n_features": len(feature_cols),
            "n_train": len(X_train),
            "n_test": len(X_test),
            "target_column": target_col,
        }

        print(f"\n--- Model Metrics (target: {target_col}) ---")
        for k, v in model_metrics.items():
            print(f"  {k}: {v}")

    if dry_run:
        print("\n[DRY RUN] Skipping Azure ML connection. Metrics computed locally.")
        print(json.dumps({"stats": stats, "model_metrics": model_metrics}, indent=2))
        return

    subscription_id = os.environ.get("AZURE_SUBSCRIPTION_ID")
    resource_group = os.environ.get("AZURE_RESOURCE_GROUP")
    workspace_name = os.environ.get("AZURE_ML_WORKSPACE_NAME")

    if not all([subscription_id, resource_group, workspace_name]):
        print("\n[SKIP] Azure ML environment variables not set.")
        print("Set AZURE_SUBSCRIPTION_ID, AZURE_RESOURCE_GROUP, AZURE_ML_WORKSPACE_NAME")
        print("Falling back to local-only mode.")
        print(json.dumps({"stats": stats, "model_metrics": model_metrics}, indent=2))
        return

    try:
        from azure.ai.ml import MLClient
        from azure.identity import DefaultAzureCredential
        from azure.ai.ml.entities import Data
        from azure.ai.ml.constants import AssetTypes
        import mlflow
    except ImportError:
        print("\n[ERROR] azure-ai-ml or mlflow not installed.")
        print("Run: pip install azure-ai-ml azure-identity mlflow")
        sys.exit(1)

    credential = DefaultAzureCredential()
    ml_client = MLClient(credential, subscription_id, resource_group, workspace_name)
    print(f"\nConnected to Azure ML workspace: {ml_client.workspace_name}")

    mlflow_tracking_uri = ml_client.workspaces.get(workspace_name).mlflow_tracking_uri
    mlflow.set_tracking_uri(mlflow_tracking_uri)

    experiment_name = "phisphere-lab-analysis"
    mlflow.set_experiment(experiment_name)

    with mlflow.start_run(run_name="phisphere-session-analysis") as run:
        mlflow.log_param("source", "phisphere-ai-export")
        mlflow.log_param("rows", len(df))
        mlflow.log_param("columns", len(df.columns))
        mlflow.log_param("column_names", ",".join(df.columns.tolist()))

        for col, s in stats.items():
            mlflow.log_metric(f"{col}_mean", s["mean"])
            mlflow.log_metric(f"{col}_std", s["std"])

        for k, v in model_metrics.items():
            if isinstance(v, (int, float)):
                mlflow.log_metric(k, v)

        if csv_path and Path(csv_path).exists():
            mlflow.log_artifact(csv_path)

        print(f"\nExperiment logged to Azure ML: {experiment_name}")
        print(f"Run ID: {run.info.run_id}")
        print(f"Tracking URI: {mlflow_tracking_uri}")

    print("\nDone. View results in Azure ML Studio.")


if __name__ == "__main__":
    main()
